requirements.txt
asgiref==3.7.2
astroid==3.0.2
autopep8==2.0.4
colorama==0.4.6
dill==0.3.7
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
isort==5.13.2
mccabe==0.7.0
pep8==1.7.1
pillow==10.2.0
platformdirs==4.1.0
pycodestyle==2.11.1
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
